requirements.txt
pydantic>=2.0
httpx>=0.27
openpyxl>=3.1
python-dotenv>=1.0
pyyaml>=6.0
pytest>=8.0
